requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
model-bakery==1.10.1
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pytest-django==4.5.2
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
types-requests==2.28.11.15
types-urllib3==1.26.25.8
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
